+++ src/main/kotlin/domain/rule/DetailRule.kt
@@ -1,0 +1,9 @@
+package domain.rule
+
+import domain.Stone
+import domain.state.State
+
+interface DetailRule {
+
+    fun stateWillObeyThisRule(state: State, nextStone: Stone): Boolean
+}
