+++ src/main/kotlin/domain/rule/RenjuRule.kt
@@ -1,0 +1,13 @@
+package domain.rule
+
+import domain.*
+import domain.state.State
+
+object RenjuRule {
+
+    private val detailRules =
+        listOf(BlackStonesAreNotAllowed33Rule, BlackStonesAreNotAllowed44Rule, BlackStonesAreNotAllowedLongMok)
+
+    fun stateWillObeyThisRule(state: State, nextStone: Stone): Boolean =
+        detailRules.all { it.stateWillObeyThisRule(state, nextStone) }
+}
